# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: sandboxlab
  version: "1.0"
  description: >-
    Create disposable sandbox environments — a shell, a filesystem, a browser —
    on demand, in Kubernetes.
servers:
  - url: https://sandboxlab.example.com/sandbox
    description: >-
      A deployment behind a gateway. The base path is the deployment's own and
      is part of the address, not something a client appends.

# Both header styles are accepted on every authenticated route, and both are
# listed so a client can use whichever it has.
security:
  - ApiKey: []
  - BearerAuth: []
...
paths:
...
  /api/v1/catalog:
    get:
      tags: [sandboxes]
      operationId: listCatalog
      summary: The templates a sandbox can be created from
      responses:
        "200":
          description: every template
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TemplateList"}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /api/v1/catalog/{id}:
    get:
      tags: [sandboxes]
      operationId: getCatalogEntry
      summary: One template
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: the template
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Template"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/overview:
    get:
      tags: [sandboxes]
      operationId: getOverview
      summary: Counts of sandboxes by state and template
      description: >-
        What is running and what it is: how many sandboxes exist, and how they
        break down by state and template.
      responses:
        "200":
          description: the summary
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Overview"}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /api/v1/sandboxes:
    get:
      tags: [sandboxes]
      operationId: listSandboxes
      summary: Every sandbox in the deployment
      responses:
        "200":
          description: the sandboxes
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SandboxList"}
        "401": {$ref: "#/components/responses/Unauthorized"}
    post:
      tags: [sandboxes]
      operationId: createSandbox
      summary: Create one
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateSandboxRequest"}
      responses:
        "201":
          description: the sandbox, which is Pending until its container is up
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Sandbox"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409": {$ref: "#/components/responses/Conflict"}
        "429": {$ref: "#/components/responses/LimitReached"}
  /api/v1/sandboxes/{id}:
    get:
      tags: [sandboxes]
      operationId: getSandbox
      summary: One sandbox
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: the sandbox
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Sandbox"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
    delete:
      tags: [sandboxes]
      operationId: deleteSandbox
      summary: Delete one
      description: >-
        Everything in it goes too. A sandbox is a namespace, so this is one
        operation rather than a list of objects to get right.
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: it was deleted
          content:
            application/json:
              schema:
                type: object
                required: [deleted]
                properties:
                  deleted:
                    type: string
                    description: the id that was deleted
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/sandboxes/{id}/renew:
    post:
      tags: [sandboxes]
      operationId: renewSandbox
      summary: Reset a sandbox's lifetime, measured from now
      description: >-
        A ttl of "0" removes the expiry entirely, which is how a sandbox someone
        is working in is kept from disappearing.
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RenewSandboxRequest"}
      responses:
        "200":
          description: the sandbox, with its new expiry
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Sandbox"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/sandboxes/{id}/logs:
    get:
      tags: [sandboxes]
      operationId: getSandboxLogs
      summary: The tail of a sandbox's output
      description: >-
        A sandbox whose container is still being created answers 409: there is
        nothing to read yet, which is the ordinary state moments after a create
        rather than a failure.
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
        - name: tail
          in: query
          required: false
          description: how many lines to return; 200 when omitted
          schema: {type: integer}
      responses:
        "200":
          description: the output
          content:
            application/json:
              schema:
                type: object
                required: [logs]
                properties:
                  logs: {type: string}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /sandbox/{id}/{port}/:
    get:
      tags: [sandboxes]
      operationId: proxyToSandbox
      summary: Proxy to a sandbox's own port
      description: >-
        The data plane: everything under this prefix is forwarded to the
        sandbox's container. What comes back is whatever the sandbox serves —
        HTML, JSON, a file — so it is not a typed endpoint and the SDK exposes
        the address rather than a method. A browser navigation cannot set a
        header, which is why the key is also accepted in the query string.
      security:
        - ApiKeyQuery: []
        - ApiKey: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
        - name: port
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: whatever the sandbox served
          content:
            "*/*":
              schema: {type: string, format: binary}
        "404": {$ref: "#/components/responses/NotFound"}
        "502":
          description: the sandbox could not be reached
components:
  securitySchemes:
    # The deployment's key. Both headers carry the same
    # credential; a client sends whichever it has.
    ApiKey:
      type: apiKey
      in: header
      name: X-Sandbox-Key
    BearerAuth:
      type: http
      scheme: bearer
    # Only the data plane takes this, because only the data plane is reached by
    # a navigation rather than by a client that can set a header.
    ApiKeyQuery:
      type: apiKey
      in: query
      name: key
  responses:
    BadRequest:
      description: the request was not usable
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    Unauthorized:
      description: no key, or not a key this deployment issued
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
...
    NotFound:
      description: no such object
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    Conflict:
      description: the request conflicts with what exists
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    LimitReached:
      description: the caller, or the deployment, is at its ceiling
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: what went wrong, in a sentence meant for the person who made the request
...
    Template:
      type: object
      required: [id, title, image]
      properties:
        id:
          type: string
          description: the stable identifier a create names
          x-oapi-codegen-extra-tags: {yaml: "id"}
        title:
          type: string
          x-oapi-codegen-extra-tags: {yaml: "title"}
        description:
          type: string
          x-oapi-codegen-extra-tags: {yaml: "description,omitempty"}
        image:
          type: string
          x-oapi-codegen-extra-tags: {yaml: "image"}
        ports:
          type: array
          items: {$ref: "#/components/schemas/Port"}
          x-oapi-codegen-extra-tags: {yaml: "ports,omitempty"}
        env:
          type: object
          additionalProperties: {type: string}
          description: environment the template sets, merged under the caller's own
          x-oapi-codegen-extra-tags: {yaml: "env,omitempty"}
        command:
          type: array
          items: {type: string}
          x-oapi-codegen-extra-tags: {yaml: "command,omitempty"}
        args:
          type: array
          items: {type: string}
          x-oapi-codegen-extra-tags: {yaml: "args,omitempty"}
        resources:
          allOf: [{$ref: "#/components/schemas/Resources"}]
          x-oapi-codegen-extra-tags: {yaml: "resources,omitempty"}
        ttlDefault:
          type: string
          description: the lifetime a create gets when it asks for none
          x-oapi-codegen-extra-tags: {yaml: "ttlDefault,omitempty"}
        ttlMax:
          type: string
          description: the ceiling a caller cannot exceed
          x-oapi-codegen-extra-tags: {yaml: "ttlMax,omitempty"}
        persistent:
          type: boolean
          x-oapi-codegen-extra-tags: {yaml: "persistent,omitempty"}
        workDir:
          type: string
          x-oapi-codegen-extra-tags: {yaml: "workDir,omitempty"}
...
    TemplateList:
      type: object
      description: The catalog — what a sandbox can be created from.
      required: [templates]
      properties:
        templates: {type: array, items: {$ref: "#/components/schemas/Template"}}

    # ── sandboxes ─────────────────────────────────────────────────────────────
...
    Sandbox:
      type: object
      required: [id, template, image, state, createdAt]
      properties:
        id:
          type: string
          description: both the name and the identity; it is a namespace name
        template: {type: string, description: "the catalog id this was created from"}
        image:    {type: string}
        state:    {$ref: "#/components/schemas/SandboxState"}
        message:
          type: string
          description: why the state is not Running, when the cluster says
        createdAt: {type: string, format: date-time}
        expiresAt:
          type: string
          format: date-time
          description: >-
            absent when the sandbox has no TTL, which is a real state rather
            than a zero time — a client that shows "no expiry" keys off its
            absence
          x-go-type: "*time.Time"
        endpoints:
          type: array
          items: {$ref: "#/components/schemas/Endpoint"}
        env:
          type: object
          additionalProperties: {type: string}
          description: the non-secret environment the sandbox was created with
        namespace: {type: string, description: "the namespace the sandbox owns"}
...
    SandboxList:
      type: object
      description: |
        Every sandbox in the deployment, and how many that is.

        Counted as well as listed because the two are not always the same
        question: a console renders the list, and a script that only wants to
        know whether anything is running should not have to read one.
      required: [sandboxes, count]
      properties:
        sandboxes: {type: array, items: {$ref: "#/components/schemas/Sandbox"}}
        count:     {type: integer}
    CreateSandboxRequest:
      type: object
      required: [template]
      properties:
        template: {type: string, description: "the catalog id to create from"}
        name:
          type: string
          description: >-
            the sandbox's name; generated when empty. Lowercase alphanumerics
            and '-', up to 40 characters, because it becomes a namespace.
        ttl: {type: string, description: "how long it may live, e.g. 90m; the template's default when empty"}
        env:
          type: object
          additionalProperties: {type: string}
          description: environment overrides, merged over the template's own
    RenewSandboxRequest:
      type: object
      required: [ttl]
      properties:
        ttl:
          type: string
          description: >-
            the new lifetime, e.g. 2h. "0" removes the expiry entirely rather
            than setting a zero one.
    Overview:
      type: object
      required: [total, byState, byTemplate, cluster, scoped]
      properties:
        total:      {type: integer}
        byState:    {type: object, additionalProperties: {type: integer}}
        byTemplate: {type: object, additionalProperties: {type: integer}}
        byOwner:    {type: object, additionalProperties: {type: integer}}
        cluster:    {type: boolean, description: "whether the cluster is reachable"}
        maxSandboxes: {type: integer, description: "the deployment's ceiling; 0 means none"}
...